Add files via upload
The current version of the class library for importing patterners generated by the Excel template in to Python. While the rest of this project is at version 1.0, having been effectively tested in a real factory in Upstate New York (the name of the client will not be disclosed), this Python Library is still in Beta at version 0.1.

Use the ModelSet constructor to create a ModelSet object representing the file in your Python environment, and call the dictionary field .patterners to find each individual obeservation of the assembly line you are studying. There are slots for up to 3 time series in each ModelSet object, and the addTimeSeries class method enables you to add them, while the .tsavailable field tells you which slots are available. Each of the three can be found as .timeseries#, while its name can be found as .tsname#.

To handle logical issues when building time series, three unique Exceptions are thrown to improve exception handling: the WhichException (which is thrown when you call a time series slot that doesn't exist), the TabException (when your time series data doesn't match the dates or times of observation as listed in the dictionary keys), and the TimeSeriesPidgeonholeException (no time series data for at least one observation).